General script cleanup
Clean up stray whitespace.
Remove the unused variables.

Drop the initial "if [ $? -ne 127 ]; then" conditional.  This appears
to do little of actual use.  It would be checking the return code from
this line:

   time="$(time ( ls ) 2>&1 1>/dev/null )"

This seems to do nothing more than check if the "ls" command exists.
That seems like a silly check.  Even if checking for the runability
of "ls" made sense, the error and exit should happen immediately after
the check.  We shouldn't wrap the entire real work in this conditional.

Next, the actual psql command.  The output from the command is saved
to a variable named archive_count_$table_name.  This variable is
unused, so we remove that.  Next the command is wrapped in _two_
different subshell mechanisms, $() and ``.  That seems unnecessarily
redundant.  But we can get rid of them both and simply run the
command directly since we don't need to capture the output in a
variable.